requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

torch==2.1.2+cu118 
torchvision==0.16.2+cu118 
torchaudio==2.1.2+cu118


transformers==4.47.0 
lm_eval==0.4.2
trl==0.12.0
peft==0.13.2
datasets==2.21.0
numpy==1.26.4
accelerate==1.0.1
wandb
huggingface_hub==0.24.5
pyyaml
python-dotenv
evaluate
tqdm
hydra-core

